ide clean up
