+++ modules/meta/arrow-meta-prototype/idea-plugin/src/main/kotlin/arrow/meta/plugin/idea/IdeMetaPlugin.kt
@@ -7,6 +7,7 @@ import arrow.meta.phases.CompilerContext
 import arrow.meta.plugin.idea.internal.registry.IdeInternalRegistry
 import arrow.meta.plugin.idea.plugins.comprehensions.comprehensionsIdePlugin
 import arrow.meta.plugin.idea.plugins.higherkinds.higherKindsIdePlugin
+import arrow.meta.plugin.idea.plugins.initial.initialIdeSetUp
 import arrow.meta.plugin.idea.plugins.nothing.nothingIdePlugin
 import arrow.meta.plugin.idea.plugins.optics.opticsIdePlugin
 import kotlin.contracts.ExperimentalContracts
@@ -15,6 +16,7 @@ class IdeMetaPlugin : MetaPlugin(), IdeInternalRegistry, IdeSyntax {
   @ExperimentalContracts
   override fun intercept(ctx: CompilerContext): List<Plugin> =
     super.intercept(ctx) +
+      initialIdeSetUp +
       nothingIdePlugin +
       comprehensionsIdePlugin +
       opticsIdePlugin +
