+++ modules/meta/arrow-meta-prototype/idea-plugin/src/main/resources/META-INF/plugin.xml
@@ -31,7 +31,6 @@
         <gradleProjectImportHandler implementation="arrow.meta.plugin.idea.gradle.ArrowGradleImportHandler"/>
         <packageFragmentProviderExtension implementation="arrow.meta.plugin.idea.phases.resolve.MetaSyntheticPackageFragmentProvider" />
 <!--        <syntheticResolveExtension implementation="arrow.meta.plugin.idea.phases.resolve.MetaSyntheticPackageFragmentProvider" />-->
-        <diagnosticSuppressor implementation="arrow.meta.plugin.idea.phases.resolve.DiagnosticSuppressor" />
     </extensions>
 
 </idea-plugin>
